Gold working with sound and board update.
When the player lands on a gold space change the board color back
to gray. This required putting the board in the db and subscribing for
board change events.  Broke the game and tile logic into separte files
because of a circluar dependency when adding it to the db.cljs file.